ENH: Fem que l'atribut que guarda el número de fase sigui un int en comptes d'un QString per ser coherents.

+++ starviewer/src/core/hangingprotocolxmlreader.cpp
@@ -302,7 +302,7 @@ HangingProtocolDisplaySet * HangingProtocolXMLReader::readDisplaySet( QXmlStream
         else if( reader->name() == "phase" )
         {
             reader->readNext();
-            displaySet->setPhase( reader->text().toString() );
+            displaySet->setPhase( reader->text().toString().toInt() );
         }
         else if( reader->name() == "displaySet" && reader->tokenType() == QXmlStreamReader::EndElement )
         {
